+++ internal/constants/constants.go
@@ -97,15 +97,3 @@ var CacheResourceTypes = []string{
 	"cluster_groups",
 	"clusters",
 }
-
-// Cable color map
-var CableColorMap = map[string]string{
-	"cat6":  "f44336",
-	"cat6a": "ffeb3b",
-	"cat7":  "ff9800",
-	"dac":   "000000",
-	"fiber": "00bcd4",
-	"om3":   "00bcd4",
-	"om4":   "2196f3",
-	"os2":   "9c27b0",
-}